Move dispatcher to public API
