+++ osquery/config/plugins/tls.cpp
@@ -15,12 +15,12 @@
 #include <boost/property_tree/ptree.hpp>
 
 #include <osquery/config.h>
+#include <osquery/dispatcher.h>
 #include <osquery/enroll.h>
 #include <osquery/flags.h>
 #include <osquery/registry.h>
 
 #include "osquery/core/conversions.h"
-#include "osquery/dispatcher/dispatcher.h"
 #include "osquery/remote/requests.h"
 #include "osquery/remote/serializers/json.h"
 #include "osquery/remote/utility.h"
