+++ osquery/dispatcher/scheduler.h
@@ -12,7 +12,7 @@
 
 #include <map>
 
-#include "osquery/dispatcher/dispatcher.h"
+#include <osquery/dispatcher.h>
 
 namespace osquery {
