+++ osquery/dispatcher/tests/dispatcher_tests.cpp
@@ -8,11 +8,9 @@
  *
  */
 
-#include <boost/make_shared.hpp>
-
 #include <gtest/gtest.h>
 
-#include "osquery/dispatcher/dispatcher.h"
+#include <osquery/dispatcher.h>
 
 namespace osquery {
