+++ osquery/events/kernel/benchmarks/kernel_communication_benchmarks.cpp
@@ -14,7 +14,8 @@
 
 #include <benchmark/benchmark.h>
 
-#include "osquery/dispatcher/dispatcher.h"
+#include <osquery/dispatcher.h>
+
 #include "osquery/events/kernel.h"
 
 namespace osquery {
